Precision matters: use double for track parameters in IO

+++ share/analysis/io.C
@@ -31,14 +31,14 @@ struct IO_t {
     int    status[kMaxTracks];
     int    parent[kMaxTracks];
     int    pdg[kMaxTracks];
-    float  vt[kMaxTracks];
-    float  vx[kMaxTracks];
-    float  vy[kMaxTracks];
-    float  vz[kMaxTracks];
-    float  e[kMaxTracks];
-    float  px[kMaxTracks];
-    float  py[kMaxTracks];
-    float  pz[kMaxTracks];
+    double vt[kMaxTracks];
+    double vx[kMaxTracks];
+    double vy[kMaxTracks];
+    double vz[kMaxTracks];
+    double  e[kMaxTracks];
+    double px[kMaxTracks];
+    double py[kMaxTracks];
+    double pz[kMaxTracks];
   } tracks;
   
   TTree *tree_hits, *tree_tracks;
